Helper Functions

Helper Functions

NeuralVerification.read_nnet — Function.

read_nnet(fname::String)

Read in neural net from file and return Network struct

source

NeuralVerification.init_layer — Function.

init_layer(i::Int64, layerSizes::Array{Int64}, f::IOStream)

Read in layer from nnet file and return a Layer struct containing its weights/biases

source

NeuralVerification.compute_output — Function.

compute_output(nnet::Network, input::Vector{Float64})

Propagate a given vector through a nnet and compute the output.

source

NeuralVerification.get_activation — Function.

get_activation(nnet::Network, x::Vector{Float64})

Given a network, find the activation pattern of all neurons at a given point x. Assume ReLU. return Vector{Vector{Bool}}. Each Vector{Bool} refers to the activation pattern of a particular layer.

source

get_activation(nnet::Network, input::Hyperrectangle)

Given a network, find the activation pattern of all neurons for a given input set. Assume ReLU. return Vector{Vector{Int64}}.

1: activated

0: undetermined

-1: not activated

source

get_activation(nnet::Network, bounds::Vector{Hyperrectangle})

Given a network, find the activation pattern of all neurons given the node-wise bounds. Assume ReLU. return Vector{Vector{Int64}}.

1: activated

0: undetermined

-1: not activated

source

NeuralVerification.get_gradient — Function.

get_gradient(nnet::Network, x::Vector)

Given a network, find the gradient at the input x

source

get_gradient(nnet::Network, input::AbstractPolytope)

Get lower and upper bounds on network gradient for a given input set. Return:

LG::Vector{Matrix}: lower bounds

UG::Vector{Matrix}: upper bounds

source

get_gradient(nnet::Network, LΛ::Vector{Matrix}, UΛ::Vector{Matrix})

Get lower and upper bounds on network gradient for given gradient bounds on activations Inputs:

LΛ::Vector{Matrix}: lower bounds on activation gradients

UΛ::Vector{Matrix}: upper bounds on activation gradients

Return:

LG::Vector{Matrix}: lower bounds

UG::Vector{Matrix}: upper bounds

source

get_gradient(nnet::Network, LΛ::Vector{Vector{N}}, UΛ::Vector{Vector{N}}) where N

Get lower and upper bounds on network gradient for given gradient bounds on activations Inputs:

LΛ::Vector{Vector{N}}: lower bounds on activation gradients

UΛ::Vector{Vector{N}}: upper bounds on activation gradients

Return:

LG::Vector{Matrix}: lower bounds

UG::Vector{Matrix}: upper bounds

source

NeuralVerification.act_gradient — Function.

act_gradient(act::ReLU, z_hat::Vector{N}) where N

Computing the gradient of an activation function at point z_hat. Currently only support ReLU.

source

NeuralVerification.act_gradient_bounds — Function.

act_gradient_bounds(nnet::Network, input::AbstractPolytope)

Computing the bounds on the gradient of all activation functions given an input set. Currently only support ReLU. Return:

LΛ::Vector{Matrix}: lower bounds

UΛ::Vector{Matrix}: upper bounds

source

NeuralVerification.interval_map — Function.

interval_map(W::Matrix{N}, l::Vector{N}, u::Vector{N}) where N

Simple linear mapping on intervals Inputs:

W::Matrix{N}: the linear mapping

l::Vector{N}: the lower bound

u::Vector{N}: the upper bound

Outputs:

l_new::Vector{N}: the lower bound after mapping

u_new::Vector{N}: the upper bound after mapping

source

interval_map(W::Matrix{N}, l::Vector{N}, u::Vector{N}) where N

Simple linear mapping on intervals Inputs:

W::Matrix{N}: the linear mapping

l::Matrix{N}: the lower bound

u::Matrix{N}: the upper bound

Outputs:

l_new::Matrix{N}: the lower bound after mapping

u_new::Matrix{N}: the upper bound after mapping

source

NeuralVerification.get_bounds — Function.

get_bounds(nnet::Network, input::Hyperrectangle)

This function calls maxSens to compute node-wise bounds given a input set.

Return:

bounds::Vector{Hyperrectangle}: bounds for all nodes AFTER activation. bounds[1] is the input set.

source

get_bounds(problem::Problem)

This function calls maxSens to compute node-wise bounds given a problem.

Return:

bounds::Vector{Hyperrectangle}: bounds for all nodes AFTER activation. bounds[1] is the input set.

source

get_bounds(nnet::Network, input::Hyperrectangle, act::Bool)

Compute bounds before or after activation by interval arithmetic. To be implemented.

Inputs:

nnet::Network: network

input::Hyperrectangle: input set

act::Bool: true for after activation bound; false for before activation bound

Return:

bounds::Vector{Hyperrectangle}: bounds for all nodes AFTER activation. bounds[1] is the input set.

source

NeuralVerification.linear_transformation — Function.

linear_transformation(layer::Layer, input::Hyperrectangle)

Transformation of a set considering linear mappings in a layer.

Inputs:

layer::Layer: a layer in a network

input::Hyperrectangle: input set

Return:

output::Hyperrectangle: set after transformation.

source

linear_transformation(layer::Layer, input::HPolytope)

Transformation of a set considering linear mappings in a layer.

Inputs:

layer::Layer: a layer in a network

input::HPolytope: input set

Return:

output::HPolytope: set after transformation.

source

linear_transformation(W::Matrix, input::HPolytope)

Transformation of a set considering a linear mapping.

Inputs:

W::Matrix: a linear mapping

input::HPolytope: input set

Return:

output::HPolytope: set after transformation.

source

NeuralVerification.split_interval — Function.

split_interval(dom::Hyperrectangle, index::Int64)

Split a set into two at the given index.

Inputs:

dom::Hyperrectangle: the set to be split

index: the index to split at

Return:

(left, right)::Tuple{Hyperrectangle, Hyperrectangle}: two sets after split

source