Helper Functions

NeuralVerification.read_nnet — Function

read_nnet(fname::String; last_layer_activation = Id())

Read in neural net from a .nnet file and return Network struct. The .nnet format is borrowed from NNet. The format assumes all hidden layers have ReLU activation. Keyword argument last_layer_activation sets the activation of the last layer, and defaults to Id(), (i.e. a linear output layer).

Missing docstring.

Missing docstring for NeuralVerification.init_layer. Check Documenter's build log for details.

NeuralVerification.compute_output — Function

compute_output(nnet::Network, input::Vector{Float64})

Propagate a given vector through a nnet and compute the output.

NeuralVerification.get_activation — Function

get_activation(L, x::Vector)

Finds the activation pattern of a vector x subject to the activation function given by the layer L. Returns a Vector{Bool} where true denotes the node is "active". In the sense of ReLU, this would be x[i] >= 0.

get_activation(nnet::Network, x::Vector)

Given a network, find the activation pattern of all neurons at a given point x. Returns Vector{Vector{Bool}}. Each Vector{Bool} refers to the activation pattern of a particular layer.

get_activation(nnet::Network, input::Hyperrectangle)

Given a network, find the activation pattern of all neurons for a given input set. Assume ReLU. return Vector{Vector{Int64}}.

1: activated

0: undetermined

-1: not activated

get_activation(nnet::Network, bounds::Vector{Hyperrectangle})

Given a network, find the activation pattern of all neurons given the node-wise bounds. Assume ReLU. return Vector{Vector{Int64}}.

1: activated

0: undetermined

-1: not activated

NeuralVerification.get_gradient — Function

get_gradient(nnet::Network, x::Vector)

Given a network, find the gradient at the input x

get_gradient(nnet::Network, input::AbstractPolytope)

Get lower and upper bounds on network gradient for a given input set. Return:

LG::Vector{Matrix}: lower bounds

UG::Vector{Matrix}: upper bounds

get_gradient(nnet::Network, LΛ::Vector{Matrix}, UΛ::Vector{Matrix})

Get lower and upper bounds on network gradient for given gradient bounds on activations Inputs:

LΛ::Vector{Matrix}: lower bounds on activation gradients

UΛ::Vector{Matrix}: upper bounds on activation gradients

Return:

LG::Vector{Matrix}: lower bounds

UG::Vector{Matrix}: upper bounds

get_gradient(nnet::Network, LΛ::Vector{Vector{N}}, UΛ::Vector{Vector{N}}) where N

Get lower and upper bounds on network gradient for given gradient bounds on activations Inputs:

LΛ::Vector{Vector{N}}: lower bounds on activation gradients

UΛ::Vector{Vector{N}}: upper bounds on activation gradients

Return:

(LG, UG) lower and upper bounds

NeuralVerification.act_gradient — Function

act_gradient(act::ReLU, z_hat::Vector{N}) where N

Computing the gradient of an activation function at point z_hat. Currently only support ReLU and Id.

NeuralVerification.act_gradient_bounds — Function

act_gradient_bounds(nnet::Network, input::AbstractPolytope)

Computing the bounds on the gradient of all activation functions given an input set. Currently only support ReLU. Return:

LΛ::Vector{Matrix}: lower bounds

UΛ::Vector{Matrix}: upper bounds

NeuralVerification.interval_map — Function

interval_map(W::Matrix, l, u)

Simple linear mapping on intervals Inputs:

W::Matrix{N}: linear mapping

l::Vector{N}: lower bound

u::Vector{N}: upper bound

Outputs:

(lbound, ubound) (after the mapping)

NeuralVerification.get_bounds — Function

get_bounds(problem::Problem)
get_bounds(nnet::Network, input::Hyperrectangle, [true])

Computes node-wise bounds given a input set. The optional last argument determines whether the bounds are pre- or post-activation.

Return:

Vector{Hyperrectangle}: bounds for all nodes. bounds[1] is the input set.

Missing docstring.

Missing docstring for NeuralVerification.linear_transformation. Check Documenter's build log for details.

NeuralVerification.split_interval — Function

split_interval(dom, i)

Split a set into two at the given index.

Inputs:

dom::Hyperrectangle: the set to be split

i: the index to split at

Return:

(left, right)::Tuple{Hyperrectangle, Hyperrectangle}: two sets after split