Problem Definitions

Problem Definitions

Problem Definitions

Problem

Network

Activation Functions

Results

Problem

NeuralVerification.Problem — Type.

Problem(network, input, output)

Problem definition for neural verification.

network is of type Network

input belongs to AbstractPolytope in LazySets.jl

output belongs to AbstractPolytope in LazySets.jl

The verification problem consists of: for all points in the input set, the corresponding output of the network must belong to the output set.

source

Network

NeuralVerification.Network — Type.

Network(layers::Vector{Layer})

Network consists of a Vector of layers.

See also: Layer

source

NeuralVerification.Layer — Type.

Layer(weights::Matrix{Float64}, bias::Vector{Float64}, activation::ActivationFunction)

Layer consists of weights and bias for linear mapping, and activation for nonlinear mapping.

See also: Network

source

NeuralVerification.Node — Type.

Node(w::Vector{Float64}, b::Float64, activation::ActivationFunction)

Node consists of w and b for linear mapping, and activation for nonlinear mapping.

source

NeuralVerification.n_nodes — Method.

n_node(L::Layer)

Returns the number of neurons in a layer.

source

Activation Functions

Note that of the activation functions listed below, only ReLU is fully supported throughout the library.

NeuralVerification.GeneralAct — Type.

GeneralAct <: ActivationFunction

Wrapper type for a general activation function.

Usage

act = GeneralAct(tanh)

act(0) == tanh(0)           # true
act(10.0) == tanh(10.0)     # true

act = GeneralAct(x->tanh.(x))

julia> act(-2:2)
5-element Array{Float64,1}:
 -0.9640275800758169
 -0.7615941559557649
  0.0
  0.7615941559557649
  0.9640275800758169

source

NeuralVerification.PiecewiseLinear — Type.

PiecewiseLinear <: ActivationFunction

Activation function that uses linear interpolation between supplied knots. An extrapolation condition can be set for values outside the set of knots. Default is Linear.

PiecewiseLinear(knots_x, knots_y, [extrapolation = Line()])

Usage

kx = [0.0, 1.2, 1.7, 3.1]
ky = [0.0, 0.5, 1.0, 1.5]
act = PiecewiseLinear(kx, ky)

act(first(kx)) == first(ky) == 0.0
act(last(kx))  == last(ky)  == 1.5

act(1.0)    # 0.4166666666666667
act(-102)   # -42.5

act = PiecewiseLinear(kx, ky, Flat())

act(-102)   # 0.0
act(Inf)    # 1.5

Extrapolations

Flat()

Line()

constant (supply a number as the argument)

Throw() (throws bounds error)

PiecewiseLinear uses Interpolations.jl.

source

NeuralVerification.Id — Type.

Id <: ActivationFunction

Identity operator

(Id())(x) -> x

source

NeuralVerification.Max — Type.

Max <: ActivationFunction

(Max())(x) -> max(maximum(x), 0)

source

NeuralVerification.ReLU — Type.

ReLU <: ActivationFunction

(ReLU())(x) -> max.(x, 0)

source

Results

NeuralVerification.Result — Type.

Result

Supertype of all result types.

See also: BasicResult, CounterExampleResult, AdversarialResult, ReachabilityResult

source

NeuralVerification.BasicResult — Type.

BasicResult(status::Symbol)

Result type that captures whether the input-output constraint is satisfied. Possible status values:

:holds (io constraint is satisfied always)

:violated (io constraint is violated)

:Unknown (could not be determined)

source

NeuralVerification.CounterExampleResult — Type.

CounterExampleResult(status, counter_example)

Like BasicResult, but also returns a counter_example if one is found (if status = :violated). The counter_example is a point in the input set that, after the NN, lies outside the output set.

source

NeuralVerification.AdversarialResult — Type.

AdversarialResult(status, max_disturbance)

Like BasicResult, but also returns the maximum allowable disturbance in the input (if status = :violated).

source

NeuralVerification.ReachabilityResult — Type.

ReachabilityResult(status, reachable)

Like BasicResult, but also returns the output reachable set given the input constraint (if status = :violated).

source